[solrbuilder] solr_builder use FnToCLI + dissolve solr_builder_main
