+++ scripts/solr_builder/setup.py
@@ -3,6 +3,6 @@ from Cython.Build import cythonize
 
 setup(
     ext_modules=cythonize(
-        "solr_builder/solr_builder_main.py",
+        "solr_builder/solr_builder.py",
         compiler_directives={'language_level': "3"})
 )
